requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
idna==2.10
numpy==1.19.4
oauthlib==3.1.0
pandas==1.1.5
Pillow==8.0.1
prettytable==2.0.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.1
tabulate==0.8.7
text-to-image==0.0.5
tweepy==3.9.0
tzlocal==2.1
urllib3==1.26.2
wcwidth==0.2.5
